pyproject.toml
[tool.poetry]
name = "checkbooknyc"
version = "0.1.0"
description = ""
authors = ["sptkl <caobaiyue@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = {version = "^3.7.4", extras = ["speedups"]} 
xmltodict = "^0.12.0"
SQLAlchemy = "^1.4.15"
psycopg2 = "^2.8.6"
python-dotenv = "^0.17.1"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
